requirements.txt
torch
diffusers
protobuf==3.20.3
python-dotenv
